portal/requirements.txt
Flask==3.0.0
gunicorn==22.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.3
Flask-WTF==1.2.1